fix unnecessary type conversions

+++ common/internal/decode/convert/convert.go
@@ -87,10 +87,10 @@ func (r *Resampling) src(i int64) (float64, float64, error) {
 	if i < 0 {
 		return 0, 0, nil
 	}
-	if r.size/4 <= int64(i) {
+	if r.size/4 <= i {
 		return 0, 0, nil
 	}
-	nextPos := int64(i) / resamplingBufferSize
+	nextPos := i / resamplingBufferSize
 	if _, ok := r.srcBufL[nextPos]; !ok {
 		if r.srcBlock+1 != nextPos {
 			if _, err := r.source.Seek(nextPos*resamplingBufferSize*4, io.SeekStart); err != nil {
